implementation of path search and walk
this is the core of the module. Path search is the coroutines that seek
for a path to given targets.
The only current implementation is a greedy one, so do not expects
efficiency in case of big graphs and few targets.

The path walk is a function that use the path search coroutine in order
to (1) prove that a path exists or (2) found the converted value for a
set of targets knowing the seeds values in an hypergraph of function.